Moved auth_token to seperate file
